# api/src/docs/openapi.yml
openapi: 3.0.0
info:
  title: Trading API
  version: 1.0.0
  description: API for managing users, balances, stock orders, and trades.
paths:
...
  /balances/inr/{userId}:
    get:
      summary: Get INR balance for a specific user.
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns INR balance for a specific user.
        400:
          description: User does not exist.
...
  /orderbook/{stockSymbol}:
    get:
      summary: Get order book for a specific stock symbol.
      parameters:
        - in: path
          name: stockSymbol
          required: true
          schema:
            type: string
      responses:
        200:
          description: Returns order book for a specific stock symbol.
        404:
          description: Order book not found for the symbol.
  /symbol/create/{stockSymbol}:
    post:
      summary: Create a new stock symbol in the order book.
      parameters:
        - in: path
          name: stockSymbol
          required: true
          schema:
            type: string
      responses:
        201:
          description: Symbol created successfully.
        400:
          description: Error creating symbol.
...